Refactor: TodoSideTab 공용으로 사용할 수 있게 변경
