+++ src/components/body/TodoSideTab.tsx
@@ -2,22 +2,11 @@ import React from 'react';
 import styled from 'styled-components';
 
 interface TodoSideTabProps {
-  taskName: string;
-  status: string;
-  createdAt: string;
+  children: React.ReactElement;
 }
 
-function TodoSideTab({ taskName, status, createdAt }: TodoSideTabProps) {
-  return (
-    <SideTabWrapper>
-      <SideTabTitle>
-        {status} {taskName}
-      </SideTabTitle>
-      <div>생성일 : {createdAt}</div>
-      <LongButton>EDIT</LongButton>
-      <RedButton>DELETE</RedButton>
-    </SideTabWrapper>
-  );
+function TodoSideTab({ children }: TodoSideTabProps): React.ReactElement {
+  return <SideTabWrapper>{children}</SideTabWrapper>;
 }
 
 export default TodoSideTab;
@@ -31,27 +20,5 @@ const SideTabWrapper = styled.div`
   right: 0;
   background-color: #fff;
   box-shadow: 0rem 0.3rem 0.9rem -0.6rem #0000005b;
-  padding: 2rem;
   z-index: 999;
-`;
-
-const SideTabTitle = styled.div`
-  font-size: 2rem;
-  font-weight: 600;
-  margin-bottom: 2rem;
-`;
-
-const LongButton = styled.button`
-  width: 100%;
-  border: 1px solid #dcdcdc;
-  border-radius: 0.8rem;
-  cursor: pointer;
-  margin-top: 0.8rem;
-  padding: 0.8rem;
-`;
-
-const RedButton = styled(LongButton)`
-  background-color: #eb7e7e;
-  color: #fff;
-  border: 1px solid #ca6d6d;
 `;
